+++ src/test/configfile_test.ts
@@ -26,12 +26,14 @@ const defaultBrowser = {
   },
 };
 
+const configFilePath = path.join(testData, 'mylib', 'tachometer.json');
+
 suite('config', () => {
   suite('parseConfigFile', () => {
     let prevCwd: string;
     suiteSetup(() => {
       prevCwd = process.cwd();
-      process.chdir(path.join(testData, 'mylib'));
+      process.chdir(path.dirname(configFilePath));
     });
 
     suiteTeardown(() => {
@@ -160,7 +162,7 @@ suite('config', () => {
           },
         ],
       };
-      const actual = await parseConfigFile(config);
+      const actual = await parseConfigFile(config, configFilePath);
       assert.deepEqual(actual, expected);
     });
 
@@ -212,7 +214,7 @@ suite('config', () => {
           },
         ],
       };
-      const actual = await parseConfigFile(config);
+      const actual = await parseConfigFile(config, configFilePath);
       assert.deepEqual(actual, expected);
     });
 
@@ -294,7 +296,7 @@ suite('config', () => {
           },
         ],
       };
-      const actual = await parseConfigFile(config);
+      const actual = await parseConfigFile(config, configFilePath);
       assert.deepEqual(actual, expected);
     });
 
@@ -377,7 +379,7 @@ suite('config', () => {
           },
         ],
       };
-      const actual = await parseConfigFile(config);
+      const actual = await parseConfigFile(config, configFilePath);
       assert.deepEqual(actual, expected);
     });
 
@@ -428,7 +430,7 @@ suite('config', () => {
           },
         ],
       };
-      const actual = await parseConfigFile(config);
+      const actual = await parseConfigFile(config, configFilePath);
       assert.deepEqual(actual, expected);
     });
 
@@ -504,7 +506,7 @@ suite('config', () => {
           },
         ],
       };
-      const actual = await parseConfigFile(config);
+      const actual = await parseConfigFile(config, configFilePath);
       assert.deepEqual(actual, expected);
     });
 
@@ -592,7 +594,7 @@ suite('config', () => {
           },
         ],
       };
-      const actual = await parseConfigFile(config);
+      const actual = await parseConfigFile(config, configFilePath);
       assert.deepEqual(actual, expected);
     });
 
@@ -600,7 +602,7 @@ suite('config', () => {
       test('invalid top-level type', async () => {
         const config = 42;
         await assert.isRejected(
-          parseConfigFile(config),
+          parseConfigFile(config, configFilePath),
           'config is not of a type(s) object'
         );
       });
@@ -610,7 +612,7 @@ suite('config', () => {
           benchmarks: 42,
         };
         await assert.isRejected(
-          parseConfigFile(config),
+          parseConfigFile(config, configFilePath),
           'config.benchmarks is not of a type(s) array'
         );
       });
@@ -620,7 +622,7 @@ suite('config', () => {
           benchmarks: [42],
         };
         await assert.isRejected(
-          parseConfigFile(config),
+          parseConfigFile(config, configFilePath),
           'config.benchmarks[0] is not of a type(s) object'
         );
       });
@@ -630,7 +632,7 @@ suite('config', () => {
           benchmarks: [],
         };
         await assert.isRejected(
-          parseConfigFile(config),
+          parseConfigFile(config, configFilePath),
           'config.benchmarks does not meet minimum length of 1'
         );
       });
@@ -640,7 +642,7 @@ suite('config', () => {
           benchmarks: [{expand: 42}],
         };
         await assert.isRejected(
-          parseConfigFile(config),
+          parseConfigFile(config, configFilePath),
           'config.benchmarks[0].expand is not of a type(s) array'
         );
       });
@@ -655,7 +657,7 @@ suite('config', () => {
           ],
         };
         await assert.isRejected(
-          parseConfigFile(config),
+          parseConfigFile(config, configFilePath),
           'config is not allowed to have the additional property "nonsense"'
         );
       });
@@ -670,7 +672,7 @@ suite('config', () => {
           ],
         };
         await assert.isRejected(
-          parseConfigFile(config),
+          parseConfigFile(config, configFilePath),
           'config.benchmarks[0] is not allowed to have the additional property "nonsense"'
         );
       });
@@ -687,7 +689,10 @@ suite('config', () => {
             },
           ],
         };
-        await assert.isRejected(parseConfigFile(config), /no url specified/i);
+        await assert.isRejected(
+          parseConfigFile(config, configFilePath),
+          /no url specified/i
+        );
       });
 
       test('unsupported browser', async () => {
@@ -700,7 +705,7 @@ suite('config', () => {
           ],
         };
         await assert.isRejected(
-          parseConfigFile(config),
+          parseConfigFile(config, configFilePath),
           'Browser potato is not supported'
         );
       });
@@ -715,7 +720,7 @@ suite('config', () => {
           ],
         };
         await assert.isRejected(
-          parseConfigFile(config),
+          parseConfigFile(config, configFilePath),
           'config.benchmarks[0].measurement is not any of: callback, fcp'
         );
       });
@@ -730,7 +735,7 @@ suite('config', () => {
           ],
         };
         await assert.isRejected(
-          parseConfigFile(config),
+          parseConfigFile(config, configFilePath),
           'config.sampleSize must be greater than or equal to 2'
         );
       });
@@ -745,7 +750,7 @@ suite('config', () => {
           ],
         };
         await assert.isRejected(
-          parseConfigFile(config),
+          parseConfigFile(config, configFilePath),
           'config.sampleSize is not of a type(s) integer'
         );
       });
@@ -764,7 +769,7 @@ suite('config', () => {
           ],
         };
         await assert.isRejected(
-          parseConfigFile(config),
+          parseConfigFile(config, configFilePath),
           'config.benchmarks[0].packageVersions requires property "label"'
         );
       });
@@ -780,7 +785,7 @@ suite('config', () => {
           ],
         };
         await assert.isRejected(
-          parseConfigFile(config),
+          parseConfigFile(config, configFilePath),
           'Please use only "autoSampleConditions" and not "horizons".'
         );
       });